// WYNKWeb/src/app/Models/OPerationTheatre.ts
export class OperationTheatre {

  OTID: number;
  OTDescription: string;
  OTType: string;
  OTAddress1: string;
  OTAddress2: string = null;
  OTAddress3: string = null;
  OTLocation: string = null;
  OTCharge: number;
 // OTDiscount: number;
  GST: number ;
  CGST: number;
  SGST: number;
  IsActive: boolean;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  CMPID: number;
}